requirements.txt
# requirements.txt

altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.5.1
attrs==23.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.8.1
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
crashtest==0.4.1
cryptography==43.0.3
cssselect==1.2.0
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
docx==0.2.4
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.0
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
installer==0.7.0
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
load-dotenv==0.1.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
numpy==1.26.4
openai==1.25.0
opensearch-py==2.5.0
packaging==24.0
pandas==2.2.2
parsel==1.9.1
pexpect==4.9.0
pillow==10.3.0
pipenv==2024.0.1
pipx==1.7.1
pkginfo==1.11.2
platformdirs==4.2.2
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
postgrest==0.16.4
Protego==0.3.1
protobuf==4.25.3
ptyprocess==0.7.0
pyarrow==16.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
PyDispatcher==2.0.7
Pygments==2.17.2
pyOpenSSL==24.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.3
queuelib==1.7.0
RapidFuzz==3.10.1
realtime==1.0.4
referencing==0.35.1
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
Scrapy==2.11.2
service-identity==24.1.0
setuptools==70.1.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
storage3==0.7.4
streamlit==1.33.0
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.5
tenacity==8.2.3
tldextract==5.1.2
toml==0.10.2
tomlkit==0.13.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.3
trove-classifiers==2024.10.21.16
Twisted==24.7.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
userpath==1.9.2
virtualenv==20.27.1
w3lib==2.2.1
watchdog==4.0.0
websockets==12.0
Werkzeug==3.0.2
zope.interface==7.1.0
